Bugfix, CDATA wrapper generates 404 on product-url
The <![CDATA[ ... ]]> wrapper causes html-entities to be read literally.
Using CDATA on a url will generate a 404 if that url contains an html-
entity.

This occurence should be very rare if friendly-urls are enabled. As
users should not be allowed to enter special characters into a
product's url. However, if the user does not have friendly-ulrs
enabled, every product-url will contain an '&' and therefore be
unnavigable to web-crawlers.