Use only the languages assigned to the shop
In the 53122f646f4c1b6889f2c3297abf2109d0e9d54a commit the $id_shop
parameter used to query languages, but if that was not specified, it
generated the list in all languages.

Use the context store ID instead: if the parameter is specified, the
context follows it.