Creates the sitemap only for the selected shop in multishop sites.
The method `createSitemap($id_shop = 0)` takes a parameter indicating
the current shop which is never actually passed. As a result of that,
the module ignores the selected shop in the admin toolbar.

To further complicate things, if you have a multistore deployement with
2 shops and 4 langs, the 2 1st langs enabled for shop 1, the remaining 2
for shop 2, creating the sitemap would create the sub-sitemaps for all
the 4 langs, which clearly is not the expected outcome.

This patch solves the problem and passes the selected shop to the
`createSitemap` method.

Co-Authored-By: Progi1984 <progi1984@gmail.com>